Commit made by the Bioconductor Git-SVN bridge.
Commit id: c4c9b898ecec446934d6fc4f05bb18bf198e29a4

    fix for lrand48



git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/Rhtslib@103494 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ src/htslib/htslib/ksort.h
@@ -66,7 +66,7 @@
 
 #ifdef __MINGW32__
 #define drand48() (rand()*(1./RAND_MAX))
-#define lrand48() (rand())
+#define lrand48() ((long int)rand())
 #define srand48(seed) (srand((seed)))
 #endif
